chore: Move FAQ into components and reroute FAQ components
