+++ components/faq.tsx
@@ -2,9 +2,9 @@ import { ChevronUpIcon } from '@heroicons/react/solid';
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
 import React, { useState, useEffect } from 'react';
-import AboutHeader from '../../components/AboutHeader';
-import FaqDisclosure from '../../components/FaqDisclosure';
-import { RequestHelper } from '../../lib/request-helper';
+import AboutHeader from './AboutHeader';
+import FaqDisclosure from './FaqDisclosure';
+import { RequestHelper } from '../lib/request-helper';
 
 /**
  * The FAQ page.
