+++ styles/globals.css
@@ -124,9 +124,9 @@ a.link {
   position: relative;
   width: 12rem;
   -webkit-transition: width 0.15s ease-in-out;
-    -moz-transition: width 0.15s ease-in-out;
-    -o-transition: width 0.15s ease-in-out;
-    transition: width 0.15s ease-in-out;
+  -moz-transition: width 0.15s ease-in-out;
+  -o-transition: width 0.15s ease-in-out;
+  transition: width 0.15s ease-in-out;
 }
 
 .menuactive {
